refactor(components): update components architecture
This moves all components to their own folder for consistency.
