+++ src/components/Hits/__tests__/Hits-test.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { shallow, mount } from 'enzyme';
 import Hits from '../Hits';
-import Template from '../../Template';
+import Template from '../../Template/Template';
 import { highlight } from '../../../helpers';
 
 describe('Hits', () => {
