+++ src/components/MenuSelect/MenuSelect.js
@@ -1,8 +1,7 @@
 import React, { Component } from 'preact-compat';
 import PropTypes from 'prop-types';
 import cx from 'classnames';
-
-import Template from './Template';
+import Template from '../Template/Template';
 
 class MenuSelect extends Component {
   static propTypes = {
