+++ src/components/RefinementList/RefinementListItem.js
@@ -1,7 +1,7 @@
 import PropTypes from 'prop-types';
 import React, { Component } from 'preact-compat';
 import isEqual from 'lodash/isEqual';
-import Template from '../Template.js';
+import Template from '../Template/Template';
 
 class RefinementListItem extends Component {
   componentWillMount() {
