+++ src/components/Selector/__tests__/Selector-test.js
