+++ src/components/Stats/Stats.js
@@ -1,6 +1,6 @@
 import React from 'preact-compat';
 import PropTypes from 'prop-types';
-import Template from '../Template.js';
+import Template from '../Template/Template';
 
 const Stats = ({
   nbHits,
