+++ src/components/Template/Template.js
@@ -1,7 +1,7 @@
 import React, { Component } from 'preact-compat';
 import PropTypes from 'prop-types';
 import isEqual from 'lodash/isEqual';
-import { isReactElement, renderTemplate } from '../lib/utils';
+import { isReactElement, renderTemplate } from '../../lib/utils';
 
 class Template extends Component {
   shouldComponentUpdate(nextProps) {
