+++ src/widgets/hits-per-page/hits-per-page.js
@@ -1,7 +1,7 @@
 import React, { render, unmountComponentAtNode } from 'preact-compat';
 import cx from 'classnames';
 import find from 'lodash/find';
-import Selector from '../../components/Selector.js';
+import Selector from '../../components/Selector/Selector.js';
 import connectHitsPerPage from '../../connectors/hits-per-page/connectHitsPerPage.js';
 import { getContainerNode } from '../../lib/utils.js';
 import { component } from '../../lib/suit.js';
